Initialize empty `Fields` slice for nil engines, extract `copyText` helper
- Return `[]` instead of `null` for `fields` in `States()` and
  `GetState()` when engine is nil, keeping JSON schema stable
- Extract shared `copyText()` clipboard helper to `clipboard.js`
- Replace inline clipboard logic in `enrol-card.jsx` and remove
  duplicate from `status-bar.jsx`
