+++ internal/web/static/app.js
@@ -1,1 +1,1 @@
-(()=>{var X,w,me,st,H,_e,he,ve,ge,ne,Z,ee,at,B={},q=[],lt=/acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i,z=Array.isArray;function F(e,t){for(var n in t)e[n]=t[n];return e}function re(e){e&&e.parentNode&&e.parentNode.removeChild(e)}function r(e,t,n){var s,l,o,i={};for(o in t)o=="key"?s=t[o]:o=="ref"?l=t[o]:i[o]=t[o];if(arguments.length>2&&(i.children=arguments.length>3?X.call(arguments,2):n),typeof e=="function"&&e.defaultProps!=null)for(o in e.defaultProps)i[o]===void 0&&(i[o]=e.defaultProps[o]);return W(e,i,s,l,null)}function W(e,t,n,s,l){var o={type:e,props:t,key:n,ref:s,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:l??++me,__i:-1,__u:0};return l==null&&w.vnode!=null&&w.vnode(o),o}function $(e){return e.children}function j(e,t){this.props=e,this.context=t}function O(e,t){if(t==null)return e.__?O(e.__,e.__i+1):null;for(var n;t<e.__k.length;t++)if((n=e.__k[t])!=null&&n.__e!=null)return n.__e;return typeof e.type=="function"?O(e):null}function it(e){if(e.__P&&e.__d){var t=e.__v,n=t.__e,s=[],l=[],o=F({},t);o.__v=t.__v+1,w.vnode&&w.vnode(o),oe(e.__P,o,t,e.__n,e.__P.namespaceURI,32&t.__u?[n]:null,s,n??O(t),!!(32&t.__u),l),o.__v=t.__v,o.__.__k[o.__i]=o,Se(s,o,l),t.__e=t.__=null,o.__e!=n&&ye(o)}}function ye(e){if((e=e.__)!=null&&e.__c!=null)return e.__e=e.__c.base=null,e.__k.some(function(t){if(t!=null&&t.__e!=null)return e.__e=e.__c.base=t.__e}),ye(e)}function pe(e){(!e.__d&&(e.__d=!0)&&H.push(e)&&!J.__r++||_e!=w.debounceRendering)&&((_e=w.debounceRendering)||he)(J)}function J(){try{for(var e,t=1;H.length;)H.length>t&&H.sort(ve),e=H.shift(),t=H.length,it(e)}finally{H.length=J.__r=0}}function be(e,t,n,s,l,o,i,u,p,c,f){var a,h,_,d,v,y,m,g=s&&s.__k||q,S=t.length;for(p=ct(n,t,g,p,S),a=0;a<S;a++)(_=n.__k[a])!=null&&(h=_.__i!=-1&&g[_.__i]||B,_.__i=a,y=oe(e,_,h,l,o,i,u,p,c,f),d=_.__e,_.ref&&h.ref!=_.ref&&(h.ref&&se(h.ref,null,_),f.push(_.ref,_.__c||d,_)),v==null&&d!=null&&(v=d),(m=!!(4&_.__u))||h.__k===_.__k?p=ke(_,p,e,m):typeof _.type=="function"&&y!==void 0?p=y:d&&(p=d.nextSibling),_.__u&=-7);return n.__e=v,p}function ct(e,t,n,s,l){var o,i,u,p,c,f=n.length,a=f,h=0;for(e.__k=new Array(l),o=0;o<l;o++)(i=t[o])!=null&&typeof i!="boolean"&&typeof i!="function"?(typeof i=="string"||typeof i=="number"||typeof i=="bigint"||i.constructor==String?i=e.__k[o]=W(null,i,null,null,null):z(i)?i=e.__k[o]=W($,{children:i},null,null,null):i.constructor===void 0&&i.__b>0?i=e.__k[o]=W(i.type,i.props,i.key,i.ref?i.ref:null,i.__v):e.__k[o]=i,p=o+h,i.__=e,i.__b=e.__b+1,u=null,(c=i.__i=ut(i,n,p,a))!=-1&&(a--,(u=n[c])&&(u.__u|=2)),u==null||u.__v==null?(c==-1&&(l>f?h--:l<f&&h++),typeof i.type!="function"&&(i.__u|=4)):c!=p&&(c==p-1?h--:c==p+1?h++:(c>p?h--:h++,i.__u|=4))):e.__k[o]=null;if(a)for(o=0;o<f;o++)(u=n[o])!=null&&!(2&u.__u)&&(u.__e==s&&(s=O(u)),we(u,u));return s}function ke(e,t,n,s){var l,o;if(typeof e.type=="function"){for(l=e.__k,o=0;l&&o<l.length;o++)l[o]&&(l[o].__=e,t=ke(l[o],t,n,s));return t}e.__e!=t&&(s&&(t&&e.type&&!t.parentNode&&(t=O(e)),n.insertBefore(e.__e,t||null)),t=e.__e);do t=t&&t.nextSibling;while(t!=null&&t.nodeType==8);return t}function ut(e,t,n,s){var l,o,i,u=e.key,p=e.type,c=t[n],f=c!=null&&(2&c.__u)==0;if(c===null&&u==null||f&&u==c.key&&p==c.type)return n;if(s>(f?1:0)){for(l=n-1,o=n+1;l>=0||o<t.length;)if((c=t[i=l>=0?l--:o++])!=null&&!(2&c.__u)&&u==c.key&&p==c.type)return i}return-1}function de(e,t,n){t[0]=="-"?e.setProperty(t,n??""):e[t]=n==null?"":typeof n!="number"||lt.test(t)?n:n+"px"}function V(e,t,n,s,l){var o,i;e:if(t=="style")if(typeof n=="string")e.style.cssText=n;else{if(typeof s=="string"&&(e.style.cssText=s=""),s)for(t in s)n&&t in n||de(e.style,t,"");if(n)for(t in n)s&&n[t]==s[t]||de(e.style,t,n[t])}else if(t[0]=="o"&&t[1]=="n")o=t!=(t=t.replace(ge,"$1")),i=t.toLowerCase(),t=i in e||t=="onFocusOut"||t=="onFocusIn"?i.slice(2):t.slice(2),e.l||(e.l={}),e.l[t+o]=n,n?s?n.u=s.u:(n.u=ne,e.addEventListener(t,o?ee:Z,o)):e.removeEventListener(t,o?ee:Z,o);else{if(l=="http://www.w3.org/2000/svg")t=t.replace(/xlink(H|:h)/,"h").replace(/sName$/,"s");else if(t!="width"&&t!="height"&&t!="href"&&t!="list"&&t!="form"&&t!="tabIndex"&&t!="download"&&t!="rowSpan"&&t!="colSpan"&&t!="role"&&t!="popover"&&t in e)try{e[t]=n??"";break e}catch{}typeof n=="function"||(n==null||n===!1&&t[4]!="-"?e.removeAttribute(t):e.setAttribute(t,t=="popover"&&n==1?"":n))}}function fe(e){return function(t){if(this.l){var n=this.l[t.type+e];if(t.t==null)t.t=ne++;else if(t.t<n.u)return;return n(w.event?w.event(t):t)}}}function oe(e,t,n,s,l,o,i,u,p,c){var f,a,h,_,d,v,y,m,g,S,k,R,C,P,D,L=t.type;if(t.constructor!==void 0)return null;128&n.__u&&(p=!!(32&n.__u),o=[u=t.__e=n.__e]),(f=w.__b)&&f(t);e:if(typeof L=="function")try{if(m=t.props,g=L.prototype&&L.prototype.render,S=(f=L.contextType)&&s[f.__c],k=f?S?S.props.value:f.__:s,n.__c?y=(a=t.__c=n.__c).__=a.__E:(g?t.__c=a=new L(m,k):(t.__c=a=new j(m,k),a.constructor=L,a.render=pt),S&&S.sub(a),a.state||(a.state={}),a.__n=s,h=a.__d=!0,a.__h=[],a._sb=[]),g&&a.__s==null&&(a.__s=a.state),g&&L.getDerivedStateFromProps!=null&&(a.__s==a.state&&(a.__s=F({},a.__s)),F(a.__s,L.getDerivedStateFromProps(m,a.__s))),_=a.props,d=a.state,a.__v=t,h)g&&L.getDerivedStateFromProps==null&&a.componentWillMount!=null&&a.componentWillMount(),g&&a.componentDidMount!=null&&a.__h.push(a.componentDidMount);else{if(g&&L.getDerivedStateFromProps==null&&m!==_&&a.componentWillReceiveProps!=null&&a.componentWillReceiveProps(m,k),t.__v==n.__v||!a.__e&&a.shouldComponentUpdate!=null&&a.shouldComponentUpdate(m,a.__s,k)===!1){t.__v!=n.__v&&(a.props=m,a.state=a.__s,a.__d=!1),t.__e=n.__e,t.__k=n.__k,t.__k.some(function(A){A&&(A.__=t)}),q.push.apply(a.__h,a._sb),a._sb=[],a.__h.length&&i.push(a);break e}a.componentWillUpdate!=null&&a.componentWillUpdate(m,a.__s,k),g&&a.componentDidUpdate!=null&&a.__h.push(function(){a.componentDidUpdate(_,d,v)})}if(a.context=k,a.props=m,a.__P=e,a.__e=!1,R=w.__r,C=0,g)a.state=a.__s,a.__d=!1,R&&R(t),f=a.render(a.props,a.state,a.context),q.push.apply(a.__h,a._sb),a._sb=[];else do a.__d=!1,R&&R(t),f=a.render(a.props,a.state,a.context),a.state=a.__s;while(a.__d&&++C<25);a.state=a.__s,a.getChildContext!=null&&(s=F(F({},s),a.getChildContext())),g&&!h&&a.getSnapshotBeforeUpdate!=null&&(v=a.getSnapshotBeforeUpdate(_,d)),P=f!=null&&f.type===$&&f.key==null?Ce(f.props.children):f,u=be(e,z(P)?P:[P],t,n,s,l,o,i,u,p,c),a.base=t.__e,t.__u&=-161,a.__h.length&&i.push(a),y&&(a.__E=a.__=null)}catch(A){if(t.__v=null,p||o!=null)if(A.then){for(t.__u|=p?160:128;u&&u.nodeType==8&&u.nextSibling;)u=u.nextSibling;o[o.indexOf(u)]=null,t.__e=u}else{for(D=o.length;D--;)re(o[D]);te(t)}else t.__e=n.__e,t.__k=n.__k,A.then||te(t);w.__e(A,t,n)}else o==null&&t.__v==n.__v?(t.__k=n.__k,t.__e=n.__e):u=t.__e=_t(n.__e,t,n,s,l,o,i,p,c);return(f=w.diffed)&&f(t),128&t.__u?void 0:u}function te(e){e&&(e.__c&&(e.__c.__e=!0),e.__k&&e.__k.some(te))}function Se(e,t,n){for(var s=0;s<n.length;s++)se(n[s],n[++s],n[++s]);w.__c&&w.__c(t,e),e.some(function(l){try{e=l.__h,l.__h=[],e.some(function(o){o.call(l)})}catch(o){w.__e(o,l.__v)}})}function Ce(e){return typeof e!="object"||e==null||e.__b>0?e:z(e)?e.map(Ce):F({},e)}function _t(e,t,n,s,l,o,i,u,p){var c,f,a,h,_,d,v,y=n.props||B,m=t.props,g=t.type;if(g=="svg"?l="http://www.w3.org/2000/svg":g=="math"?l="http://www.w3.org/1998/Math/MathML":l||(l="http://www.w3.org/1999/xhtml"),o!=null){for(c=0;c<o.length;c++)if((_=o[c])&&"setAttribute"in _==!!g&&(g?_.localName==g:_.nodeType==3)){e=_,o[c]=null;break}}if(e==null){if(g==null)return document.createTextNode(m);e=document.createElementNS(l,g,m.is&&m),u&&(w.__m&&w.__m(t,o),u=!1),o=null}if(g==null)y===m||u&&e.data==m||(e.data=m);else{if(o=o&&X.call(e.childNodes),!u&&o!=null)for(y={},c=0;c<e.attributes.length;c++)y[(_=e.attributes[c]).name]=_.value;for(c in y)_=y[c],c=="dangerouslySetInnerHTML"?a=_:c=="children"||c in m||c=="value"&&"defaultValue"in m||c=="checked"&&"defaultChecked"in m||V(e,c,null,_,l);for(c in m)_=m[c],c=="children"?h=_:c=="dangerouslySetInnerHTML"?f=_:c=="value"?d=_:c=="checked"?v=_:u&&typeof _!="function"||y[c]===_||V(e,c,_,y[c],l);if(f)u||a&&(f.__html==a.__html||f.__html==e.innerHTML)||(e.innerHTML=f.__html),t.__k=[];else if(a&&(e.innerHTML=""),be(t.type=="template"?e.content:e,z(h)?h:[h],t,n,s,g=="foreignObject"?"http://www.w3.org/1999/xhtml":l,o,i,o?o[0]:n.__k&&O(n,0),u,p),o!=null)for(c=o.length;c--;)re(o[c]);u||(c="value",g=="progress"&&d==null?e.removeAttribute("value"):d!=null&&(d!==e[c]||g=="progress"&&!d||g=="option"&&d!=y[c])&&V(e,c,d,y[c],l),c="checked",v!=null&&v!=e[c]&&V(e,c,v,y[c],l))}return e}function se(e,t,n){try{if(typeof e=="function"){var s=typeof e.__u=="function";s&&e.__u(),s&&t==null||(e.__u=e(t))}else e.current=t}catch(l){w.__e(l,n)}}function we(e,t,n){var s,l;if(w.unmount&&w.unmount(e),(s=e.ref)&&(s.current&&s.current!=e.__e||se(s,null,t)),(s=e.__c)!=null){if(s.componentWillUnmount)try{s.componentWillUnmount()}catch(o){w.__e(o,t)}s.base=s.__P=null}if(s=e.__k)for(l=0;l<s.length;l++)s[l]&&we(s[l],t,n||typeof e.type!="function");n||re(e.__e),e.__c=e.__=e.__e=void 0}function pt(e,t,n){return this.constructor(e,n)}function xe(e,t,n){var s,l,o,i;t==document&&(t=document.documentElement),w.__&&w.__(e,t),l=(s=typeof n=="function")?null:n&&n.__k||t.__k,o=[],i=[],oe(t,e=(!s&&n||t).__k=r($,null,[e]),l||B,B,t.namespaceURI,!s&&n?[n]:l?null:t.firstChild?X.call(t.childNodes):null,o,!s&&n?n:l?l.__e:t.firstChild,s,i),Se(o,e,i)}X=q.slice,w={__e:function(e,t,n,s){for(var l,o,i;t=t.__;)if((l=t.__c)&&!l.__)try{if((o=l.constructor)&&o.getDerivedStateFromError!=null&&(l.setState(o.getDerivedStateFromError(e)),i=l.__d),l.componentDidCatch!=null&&(l.componentDidCatch(e,s||{}),i=l.__d),i)return l.__E=l}catch(u){e=u}throw e}},me=0,st=function(e){return e!=null&&e.constructor===void 0},j.prototype.setState=function(e,t){var n;n=this.__s!=null&&this.__s!=this.state?this.__s:this.__s=F({},this.state),typeof e=="function"&&(e=e(F({},n),this.props)),e&&F(n,e),e!=null&&this.__v&&(t&&this._sb.push(t),pe(this))},j.prototype.forceUpdate=function(e){this.__v&&(this.__e=!0,e&&this.__h.push(e),pe(this))},j.prototype.render=$,H=[],he=typeof Promise=="function"?Promise.prototype.then.bind(Promise.resolve()):setTimeout,ve=function(e,t){return e.__v.__b-t.__v.__b},J.__r=0,ge=/(PointerCapture)$|Capture$/i,ne=0,Z=fe(!1),ee=fe(!0),at=0;var M,x,ae,Te,K=0,He=[],T=w,Pe=T.__b,Ee=T.__r,Re=T.diffed,De=T.__c,Le=T.unmount,Ie=T.__;function ie(e,t){T.__h&&T.__h(x,e,K||t),K=0;var n=x.__H||(x.__H={__:[],__h:[]});return e>=n.__.length&&n.__.push({}),n.__[e]}function b(e){return K=1,dt(Ne,e)}function dt(e,t,n){var s=ie(M++,2);if(s.t=e,!s.__c&&(s.__=[n?n(t):Ne(void 0,t),function(u){var p=s.__N?s.__N[0]:s.__[0],c=s.t(p,u);p!==c&&(s.__N=[c,s.__[1]],s.__c.setState({}))}],s.__c=x,!x.__f)){var l=function(u,p,c){if(!s.__c.__H)return!0;var f=s.__c.__H.__.filter(function(h){return h.__c});if(f.every(function(h){return!h.__N}))return!o||o.call(this,u,p,c);var a=s.__c.props!==u;return f.some(function(h){if(h.__N){var _=h.__[0];h.__=h.__N,h.__N=void 0,_!==h.__[0]&&(a=!0)}}),o&&o.call(this,u,p,c)||a};x.__f=!0;var o=x.shouldComponentUpdate,i=x.componentWillUpdate;x.componentWillUpdate=function(u,p,c){if(this.__e){var f=o;o=void 0,l(u,p,c),o=f}i&&i.call(this,u,p,c)},x.shouldComponentUpdate=l}return s.__N||s.__}function I(e,t){var n=ie(M++,3);!T.__s&&Ue(n.__H,t)&&(n.__=e,n.u=t,x.__H.__h.push(n))}function N(e){return K=5,ft(function(){return{current:e}},[])}function ft(e,t){var n=ie(M++,7);return Ue(n.__H,t)&&(n.__=e(),n.__H=t,n.__h=e),n.__}function mt(){for(var e;e=He.shift();){var t=e.__H;if(e.__P&&t)try{t.__h.some(G),t.__h.some(le),t.__h=[]}catch(n){t.__h=[],T.__e(n,e.__v)}}}T.__b=function(e){x=null,Pe&&Pe(e)},T.__=function(e,t){e&&t.__k&&t.__k.__m&&(e.__m=t.__k.__m),Ie&&Ie(e,t)},T.__r=function(e){Ee&&Ee(e),M=0;var t=(x=e.__c).__H;t&&(ae===x?(t.__h=[],x.__h=[],t.__.some(function(n){n.__N&&(n.__=n.__N),n.u=n.__N=void 0})):(t.__h.some(G),t.__h.some(le),t.__h=[],M=0)),ae=x},T.diffed=function(e){Re&&Re(e);var t=e.__c;t&&t.__H&&(t.__H.__h.length&&(He.push(t)!==1&&Te===T.requestAnimationFrame||((Te=T.requestAnimationFrame)||ht)(mt)),t.__H.__.some(function(n){n.u&&(n.__H=n.u),n.u=void 0})),ae=x=null},T.__c=function(e,t){t.some(function(n){try{n.__h.some(G),n.__h=n.__h.filter(function(s){return!s.__||le(s)})}catch(s){t.some(function(l){l.__h&&(l.__h=[])}),t=[],T.__e(s,n.__v)}}),De&&De(e,t)},T.unmount=function(e){Le&&Le(e);var t,n=e.__c;n&&n.__H&&(n.__H.__.some(function(s){try{G(s)}catch(l){t=l}}),n.__H=void 0,t&&T.__e(t,n.__v))};var Fe=typeof requestAnimationFrame=="function";function ht(e){var t,n=function(){clearTimeout(s),Fe&&cancelAnimationFrame(t),setTimeout(e)},s=setTimeout(n,35);Fe&&(t=requestAnimationFrame(n))}function G(e){var t=x,n=e.__c;typeof n=="function"&&(e.__c=void 0,n()),x=t}function le(e){var t=x;e.__c=e.__(),x=t}function Ue(e,t){return!e||e.length!==t.length||t.some(function(n,s){return n!==e[s]})}function Ne(e,t){return typeof t=="function"?t(e):t}var Q=null;async function E(e,t={}){let n=await fetch(e,t);if(!n.ok){let s=await n.json().catch(()=>({error:n.statusText}));throw new Error(s.error||n.statusText)}return n.json()}async function U(){Q||(Q=(await E("/api/v1/csrf")).token);let e=Q;return Q=null,e}async function ce(){return E("/api/v1/status")}async function Ae(){return E("/api/v1/rules")}async function Oe(e=""){return E(`/api/v1/secrets/${e}`)}async function Y(e,t=!1){let n=t?`/api/v1/secrets/${e}?reveal=true`:`/api/v1/secrets/${e}`;return E(n)}async function $e(){let e=await U();return E("/api/v1/sync",{method:"POST",headers:{"X-CSRF-Token":e}})}async function Me(e,t){let n=await U();return E("/auth/ldap/login",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":n},body:JSON.stringify({username:e,password:t})})}async function Ve(e){return E(`/auth/ldap/status?session=${encodeURIComponent(e)}`)}async function We(e,t){let n=await U();return E("/auth/ldap/totp",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":n},body:JSON.stringify({session_id:e,passcode:t})})}async function je(e){let t=await U();return E("/auth/token/login",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":t},body:JSON.stringify({token:e})})}async function Be(){return E("/api/v1/token")}async function qe(e){let t=await U();return E(`/api/v1/enrol/${encodeURIComponent(e)}/start`,{method:"POST",headers:{"X-CSRF-Token":t}})}async function Je(e){let t=await U();return E(`/api/v1/enrol/${encodeURIComponent(e)}/skip`,{method:"POST",headers:{"X-CSRF-Token":t}})}async function Xe(e){return E(`/api/v1/enrol/${encodeURIComponent(e)}/status`)}async function ze(){let e=await U();return E("/api/v1/enrol/complete",{method:"POST",headers:{"X-CSRF-Token":e}})}async function Ge(){return E("/api/v1/enrol/prompt")}async function Ke(e){let t=await U();return E("/api/v1/enrol/secret",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":t},body:JSON.stringify({value:e})})}function vt(e){let t=document.createElement("textarea");t.value=e,t.style.position="fixed",t.style.opacity="0",document.body.appendChild(t),t.select(),document.execCommand("copy"),document.body.removeChild(t)}function Qe({status:e,onSync:t,pendingEnrolments:n,onEnrolClick:s}){let[l,o]=b(!1),[i,u]=b(!1),p=N(null);async function c(){o(!0);try{await $e(),t&&await t()}catch(v){console.error("sync failed:",v)}finally{o(!1)}}async function f(){try{if(!p.current){let v=await Be();p.current=v.token}vt(p.current),u(!0),setTimeout(()=>u(!1),2e3)}catch(v){console.error("copy token failed:",v)}}let a=e?.authenticated?"Connected":"Disconnected",h=e?.authenticated?"status-ok":"status-error",_=e?.vault_address,d=_&&/^https?:\/\//i.test(_)?_:null;return r("header",{class:"status-bar"},r("div",{class:"status-left"},r("span",{class:"app-title"},".vault",e?.version&&r("span",{class:"app-version"}," v"+e.version)),r("span",{class:`status-indicator ${h}`},a),d&&r("a",{class:"vault-link",href:d,target:"_blank",rel:"noopener noreferrer"},"Vault \u2197"),n>0&&r("button",{class:"enrol-indicator",onClick:s},n+" pending")),r("div",{class:"status-right"},e?.time&&r("span",{class:"last-sync"},"Updated: ",new Date(e.time).toLocaleTimeString()),e?.authenticated&&r("button",{class:"copy-token-btn"+(i?" copied":""),onClick:f,title:i?"Token copied!":"Copy Vault token to clipboard"},i?"\u2705":"\u{1F4CB}"),r("button",{class:"sync-btn",onClick:c,disabled:l},l?"Syncing...":"Sync Now")))}function Ye({keys:e,selected:t,onSelect:n}){return!e||e.length===0?r("nav",{class:"sidebar"},r("div",{class:"sidebar-empty"},"No secrets found")):r("nav",{class:"sidebar"},r("div",{class:"sidebar-header"},"Secrets"),r("ul",{class:"key-list"},e.map(s=>r("li",{key:s,class:`key-item ${s===t?"selected":""}`,onClick:()=>n(s)},r("span",{class:"key-icon"},s.endsWith("/")?"\u{1F4C1}":"\u{1F511}"),r("span",{class:"key-name"},s)))))}function gt(e){return!!e&&/^https?:\/\//i.test(e)}function yt(e,t){if(!gt(e.vault_address))return null;let n=e.vault_address.replace(/\/+$/,""),s=t&&t.endsWith("/")?"list":"show",l=encodeURIComponent(e.kv_mount),o=c=>(c||"").split("/").filter(f=>f.length>0).map(f=>encodeURIComponent(f)).join("/"),i=o(`${e.user_prefix||""}${e.username}`),u=o(t),p=u?`${i}/${u}`:i;return`${n}/ui/vault/secrets/${l}/${s}/${p}`}function Ze({secretPath:e,status:t,customText:n}){let[s,l]=b(null),[o,i]=b({}),[u,p]=b({}),[c,f]=b(!1);I(()=>{if(!e){l(null);return}f(!0),i({}),Y(e).then(l).catch(()=>l(null)).finally(()=>f(!1))},[e]);async function a(d){if(o[d]){i(y=>{let m={...y};return delete m[d],m});return}let v=await Y(e,!0);v&&v.fields&&(i(y=>({...y,[d]:v.fields[d]})),setTimeout(()=>{i(y=>{let m={...y};return delete m[d],m})},3e4))}async function h(d){let v=await Y(e,!0);if(v&&v.fields){let y=typeof v.fields[d]=="object"?JSON.stringify(v.fields[d],null,2):String(v.fields[d]);await navigator.clipboard.writeText(y),p(m=>({...m,[d]:!0})),setTimeout(()=>{p(m=>{let g={...m};return delete g[d],g})},2e3)}}if(!e)return r("main",{class:"secret-panel"},n?r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:n}}):r("div",{class:"panel-empty"},"Select a secret from the sidebar"));if(c)return r("main",{class:"secret-panel"},r("div",{class:"panel-loading"},"Loading..."));if(!s)return r("main",{class:"secret-panel"},r("div",{class:"panel-error"},"Secret not found"));let _=Array.isArray(s.fields)?s.fields:Object.keys(s.fields||{});return r("main",{class:"secret-panel"},r("div",{class:"secret-heading"},r("h2",null,e),(()=>{let d=t?.kv_mount&&t?.username&&yt(t,e);return d&&r("a",{class:"vault-link",href:d,target:"_blank",rel:"noopener noreferrer"},"View in Vault \u2197")})()),r("div",{class:"secret-meta"},"Version: ",s.version),r("table",{class:"field-table"},r("thead",null,r("tr",null,r("th",null,"Field"),r("th",null,"Value"),r("th",null,""))),r("tbody",null,_.map(d=>r("tr",{key:d},r("td",{class:"field-name"},d),r("td",{class:"field-value"},o[d]?r("code",null,typeof o[d]=="object"?JSON.stringify(o[d],null,2):String(o[d])):r("span",{class:"masked"},"\u2022\u2022\u2022\u2022\u2022\u2022\u2022\u2022")),r("td",{class:"field-actions"},r("button",{class:"reveal-btn",onClick:()=>a(d),title:o[d]?"Hide":"Reveal"},o[d]?"\u{1F441}\uFE0F":"\u{1F441}"),r("button",{class:"copy-btn"+(u[d]?" copied":""),onClick:()=>h(d),title:u[d]?"Copied!":"Copy to clipboard"},u[d]?"\u2705":"\u{1F4CB}")))))))}function et({rules:e}){return!e||e.length===0?null:r("div",{class:"oauth-banners"},e.map(t=>r("div",{key:t.name,class:"oauth-banner"},r("span",null,t.oauth_provider||t.name," requires authorization"),r("a",{class:"oauth-btn",href:`/api/v1/oauth/${t.name}/start`},"Authorize"))))}function tt({authMethod:e,onAuth:t,customText:n}){return e==="oidc"?r(bt,{customText:n}):e==="ldap"?r(kt,{onAuth:t,customText:n}):e==="token"?r(St,{onAuth:t,customText:n}):r("div",{class:"login-container"},r("p",{class:"login-error"},`Unknown auth method: ${e}`))}function bt({customText:e}){return r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),e&&r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:e}}),r("a",{class:"login-btn",href:"/auth/oidc/start"},"Login with OIDC")))}function kt({onAuth:e,customText:t}){let[n,s]=b(""),[l,o]=b(""),[i,u]=b(null),[p,c]=b("form"),[f,a]=b([]),[h,_]=b(""),d=N(null),v=N(null);I(()=>()=>{v.current&&clearInterval(v.current)},[]);function y(S){v.current&&clearInterval(v.current),v.current=setInterval(async()=>{try{let k=await Ve(S);if(k.state==="mfa_required"){a(k.mfa_methods||[]);let R=k.mfa_methods&&k.mfa_methods.some(C=>C.uses_passcode);c(R?"totp":"polling"),k.error&&u(k.error)}else k.state==="authenticated"?(clearInterval(v.current),e&&e()):k.state==="failed"&&(clearInterval(v.current),u(k.error||"Login failed"),c("form"))}catch(k){clearInterval(v.current),u(k.message),c("form")}},2e3)}async function m(S){S.preventDefault(),u(null),c("polling");try{let k=await Me(n,l);d.current=k.session_id,y(k.session_id)}catch(k){u(k.message),c("form")}}async function g(S){S.preventDefault(),u(null);try{await We(d.current,h),_("")}catch(k){u(k.message)}}return p==="polling"&&(!f.length||!f.some(S=>S.uses_passcode))?r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),r("p",{class:"login-mfa-wait"},"Waiting for MFA approval..."),r("p",{class:"login-mfa-hint"},"Check your device"),i&&r("p",{class:"login-error"},i))):p==="totp"?r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),r("p",{class:"login-subtitle"},"Enter MFA passcode"),i&&r("p",{class:"login-error"},i),r("form",{onSubmit:g},r("input",{type:"text",class:"login-input",placeholder:"Passcode","aria-label":"MFA Passcode",value:h,onInput:S=>_(S.target.value),autocomplete:"one-time-code",inputMode:"numeric",autofocus:!0}),r("button",{type:"submit",class:"login-btn",disabled:!h},"Verify")))):r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),t&&r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:t}}),i&&r("p",{class:"login-error"},i),r("form",{onSubmit:m},r("input",{type:"text",class:"login-input",placeholder:"Username","aria-label":"Username",value:n,onInput:S=>s(S.target.value),autocomplete:"username",autofocus:!0}),r("input",{type:"password",class:"login-input",placeholder:"Password","aria-label":"Password",value:l,onInput:S=>o(S.target.value),autocomplete:"current-password"}),r("button",{type:"submit",class:"login-btn",disabled:!n||!l},"Login"))))}function St({onAuth:e,customText:t}){let[n,s]=b(""),[l,o]=b(null),[i,u]=b(!1);async function p(c){c.preventDefault(),o(null),u(!0);try{await je(n),e&&e()}catch(f){o(f.message)}finally{u(!1)}}return r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),t&&r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:t}}),l&&r("p",{class:"login-error"},l),r("form",{onSubmit:p},r("input",{type:"password",class:"login-input",placeholder:"Vault Token","aria-label":"Vault Token",value:n,onInput:c=>s(c.target.value),autofocus:!0}),r("button",{type:"submit",class:"login-btn",disabled:!n||i},i?"Validating...":"Login"))))}function nt({enrolment:e,onUpdate:t,anyRunning:n}){let[s,l]=b(e.status),[o,i]=b(e.output||[]),[u,p]=b(e.error||null),[c,f]=b(null),[a,h]=b(""),_=N(null);I(()=>{l(e.status),p(e.error||null),e.output&&i(e.output)},[e.status,e.error,e.output]),I(()=>(s==="running"&&!_.current&&d(),()=>{_.current&&(clearInterval(_.current),_.current=null)}),[s]);function d(){_.current&&(clearInterval(_.current),_.current=null);let C=e.engine!=="github";_.current=setInterval(async()=>{try{let P=await Xe(e.key);if(i(P.output||[]),C){let D=await Ge();D.pending?f(D.label):f(null)}P.status!=="running"&&(clearInterval(_.current),_.current=null,l(P.status),p(P.error||null),f(null),t&&t())}catch(P){console.error("poll error:",P)}},2e3)}async function v(){try{await qe(e.key),l("running"),i([]),p(null),t&&t(),d()}catch(C){p(C.message)}}async function y(){try{await Je(e.key),l("skipped"),t&&t()}catch(C){p(C.message)}}async function m(C){C.preventDefault();try{await Ke(a),h(""),f(null)}catch(P){p(P.message)}}let g=o.reduce((C,P)=>{let D=P.match(/one-time code: (\S+)/);return D?D[1]:C},null),S=o.reduce((C,P)=>{let D=P.match(/open (https?:\/\/\S+)/);return D?D[1]:C},null),k=e.engine==="github",R=n&&s!=="running";return s==="complete"?r("div",{class:"enrol-card enrol-complete"},r("div",{class:"enrol-card-header"},r("div",null,r("span",{class:"enrol-check"},"\u2713"),r("strong",null,e.name)),r("span",{class:"enrol-status-text enrol-status-complete"},"Enrolled successfully"))):s==="skipped"?r("div",{class:"enrol-card enrol-skipped"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-badge"},"SKIPPED")),r("span",{class:"enrol-engine-desc"},ue(e.engine)))):s==="running"?r("div",{class:"enrol-card enrol-running"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-badge enrol-badge-running"},"RUNNING"))),k&&g&&r("div",{class:"enrol-device-flow"},r("p",{class:"enrol-device-label"},"Enter this code on GitHub:"),r("div",{class:"enrol-device-code"},g),r("div",{class:"enrol-device-actions"},r("button",{class:"enrol-btn-secondary",onClick:()=>{let C=document.createElement("textarea");C.value=g,C.style.position="fixed",C.style.opacity="0",document.body.appendChild(C),C.select(),document.execCommand("copy"),document.body.removeChild(C)}},"Copy Code"),S&&r("a",{class:"enrol-btn-secondary",href:S,target:"_blank",rel:"noopener noreferrer"},"Open GitHub \u2192")),r("p",{class:"enrol-device-waiting"},"Waiting for approval...")),c&&!k&&r("form",{class:"enrol-prompt-form",onSubmit:m},r("label",{class:"enrol-prompt-label"},c),r("input",{type:"password",class:"enrol-prompt-input",value:a,onInput:C=>h(C.target.value),placeholder:"Enter passphrase",autofocus:!0}),r("div",{class:"enrol-prompt-actions"},r("button",{type:"submit",class:"enrol-btn-primary"},"Submit"))),!k&&!c&&o.length>0&&r("div",{class:"enrol-output"},o.map((C,P)=>r("div",{key:P},C)))):s==="failed"?r("div",{class:"enrol-card enrol-failed"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-engine-desc"},ue(e.engine))),r("div",{class:"enrol-card-actions"},r("button",{class:"enrol-btn-primary",onClick:v,disabled:R},"Retry"),r("button",{class:"enrol-btn-secondary",onClick:y,disabled:R},"Skip"))),r("p",{class:"enrol-error-text"},u)):r("div",{class:"enrol-card"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-engine-desc"},ue(e.engine))),r("div",{class:"enrol-card-actions"},r("button",{class:"enrol-btn-primary",onClick:v,disabled:R},"Start"),r("button",{class:"enrol-btn-secondary",onClick:y,disabled:R},"Skip"))),u&&r("p",{class:"enrol-error-text"},u))}function ue(e){switch(e){case"github":return"OAuth token via device flow";case"ssh":return"Ed25519 key generation";default:return e}}function rt({enrolments:e,onComplete:t,onUpdate:n}){let s=e.every(i=>i.status==="complete"||i.status==="skipped"),l=e.some(i=>i.status==="running");async function o(){try{await ze(),t&&t()}catch(i){console.error("complete error:",i)}}return r("div",{class:"enrol-page"},r("div",{class:"enrol-container"},r("h2",{class:"enrol-heading"},"Complete your setup"),r("p",{class:"enrol-subheading"},"The following credentials need to be configured before syncing can begin."),r("div",{class:"enrol-list"},e.map(i=>r(nt,{key:i.key,enrolment:i,onUpdate:n,anyRunning:l}))),r("div",{class:"enrol-footer"},r("button",{class:"enrol-continue-btn",onClick:o,disabled:l},s?"Continue to Dashboard \u2192":"Skip remaining and continue \u2192"))))}function ot(){let[e,t]=b(null),[n,s]=b([]),[l,o]=b([]),[i,u]=b(null),[p,c]=b(null),[f,a]=b(!1);I(()=>{h()},[]),I(()=>{if(!e?.authenticated)return;let m=setInterval(_,3e4);return()=>clearInterval(m)},[e?.authenticated]);async function h(){try{let m=await ce();t(m),m.authenticated&&_()}catch(m){c(m.message)}}async function _(){try{let[m,g,S]=await Promise.all([ce(),Ae(),Oe()]);t(m),s(g.rules||[]),o(S.keys||[]),c(null)}catch(m){c(m.message)}}if(e&&!e.authenticated)return r(tt,{authMethod:e.auth_method,onAuth:_,customText:e.login_text});if(!e)return r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),p?r("p",{class:"login-error"},p):r("p",null,"Loading...")));let d=e.enrolments||[],v=d.filter(m=>m.status==="pending"||m.status==="running"||m.status==="failed");if(v.length>0&&!f)return r(rt,{enrolments:d,onComplete:()=>{a(!0),_()},onUpdate:h});let y=n.filter(m=>m.has_oauth);return r($,null,r(Qe,{status:e,onSync:_,pendingEnrolments:v.length,onEnrolClick:()=>a(!1)}),p&&r("div",{class:"error-banner"},p),y.length>0&&r(et,{rules:y}),r("div",{class:"main-layout"},r(Ye,{keys:l,selected:i,onSelect:u}),r(Ze,{secretPath:i,status:e,customText:e.secret_view_text})))}xe(r(ot,null),document.getElementById("app"));})();
+(()=>{var X,C,he,at,H,pe,ve,ge,ye,re,ee,te,lt,B={},q=[],it=/acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i,z=Array.isArray;function F(e,t){for(var n in t)e[n]=t[n];return e}function oe(e){e&&e.parentNode&&e.parentNode.removeChild(e)}function r(e,t,n){var s,l,o,i={};for(o in t)o=="key"?s=t[o]:o=="ref"?l=t[o]:i[o]=t[o];if(arguments.length>2&&(i.children=arguments.length>3?X.call(arguments,2):n),typeof e=="function"&&e.defaultProps!=null)for(o in e.defaultProps)i[o]===void 0&&(i[o]=e.defaultProps[o]);return W(e,i,s,l,null)}function W(e,t,n,s,l){var o={type:e,props:t,key:n,ref:s,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:l??++he,__i:-1,__u:0};return l==null&&C.vnode!=null&&C.vnode(o),o}function $(e){return e.children}function j(e,t){this.props=e,this.context=t}function O(e,t){if(t==null)return e.__?O(e.__,e.__i+1):null;for(var n;t<e.__k.length;t++)if((n=e.__k[t])!=null&&n.__e!=null)return n.__e;return typeof e.type=="function"?O(e):null}function ct(e){if(e.__P&&e.__d){var t=e.__v,n=t.__e,s=[],l=[],o=F({},t);o.__v=t.__v+1,C.vnode&&C.vnode(o),se(e.__P,o,t,e.__n,e.__P.namespaceURI,32&t.__u?[n]:null,s,n??O(t),!!(32&t.__u),l),o.__v=t.__v,o.__.__k[o.__i]=o,Ce(s,o,l),t.__e=t.__=null,o.__e!=n&&be(o)}}function be(e){if((e=e.__)!=null&&e.__c!=null)return e.__e=e.__c.base=null,e.__k.some(function(t){if(t!=null&&t.__e!=null)return e.__e=e.__c.base=t.__e}),be(e)}function fe(e){(!e.__d&&(e.__d=!0)&&H.push(e)&&!J.__r++||pe!=C.debounceRendering)&&((pe=C.debounceRendering)||ve)(J)}function J(){try{for(var e,t=1;H.length;)H.length>t&&H.sort(ge),e=H.shift(),t=H.length,ct(e)}finally{H.length=J.__r=0}}function ke(e,t,n,s,l,o,i,u,p,c,d){var a,h,_,f,v,y,m,g=s&&s.__k||q,S=t.length;for(p=ut(n,t,g,p,S),a=0;a<S;a++)(_=n.__k[a])!=null&&(h=_.__i!=-1&&g[_.__i]||B,_.__i=a,y=se(e,_,h,l,o,i,u,p,c,d),f=_.__e,_.ref&&h.ref!=_.ref&&(h.ref&&ae(h.ref,null,_),d.push(_.ref,_.__c||f,_)),v==null&&f!=null&&(v=f),(m=!!(4&_.__u))||h.__k===_.__k?p=Se(_,p,e,m):typeof _.type=="function"&&y!==void 0?p=y:f&&(p=f.nextSibling),_.__u&=-7);return n.__e=v,p}function ut(e,t,n,s,l){var o,i,u,p,c,d=n.length,a=d,h=0;for(e.__k=new Array(l),o=0;o<l;o++)(i=t[o])!=null&&typeof i!="boolean"&&typeof i!="function"?(typeof i=="string"||typeof i=="number"||typeof i=="bigint"||i.constructor==String?i=e.__k[o]=W(null,i,null,null,null):z(i)?i=e.__k[o]=W($,{children:i},null,null,null):i.constructor===void 0&&i.__b>0?i=e.__k[o]=W(i.type,i.props,i.key,i.ref?i.ref:null,i.__v):e.__k[o]=i,p=o+h,i.__=e,i.__b=e.__b+1,u=null,(c=i.__i=_t(i,n,p,a))!=-1&&(a--,(u=n[c])&&(u.__u|=2)),u==null||u.__v==null?(c==-1&&(l>d?h--:l<d&&h++),typeof i.type!="function"&&(i.__u|=4)):c!=p&&(c==p-1?h--:c==p+1?h++:(c>p?h--:h++,i.__u|=4))):e.__k[o]=null;if(a)for(o=0;o<d;o++)(u=n[o])!=null&&!(2&u.__u)&&(u.__e==s&&(s=O(u)),xe(u,u));return s}function Se(e,t,n,s){var l,o;if(typeof e.type=="function"){for(l=e.__k,o=0;l&&o<l.length;o++)l[o]&&(l[o].__=e,t=Se(l[o],t,n,s));return t}e.__e!=t&&(s&&(t&&e.type&&!t.parentNode&&(t=O(e)),n.insertBefore(e.__e,t||null)),t=e.__e);do t=t&&t.nextSibling;while(t!=null&&t.nodeType==8);return t}function _t(e,t,n,s){var l,o,i,u=e.key,p=e.type,c=t[n],d=c!=null&&(2&c.__u)==0;if(c===null&&u==null||d&&u==c.key&&p==c.type)return n;if(s>(d?1:0)){for(l=n-1,o=n+1;l>=0||o<t.length;)if((c=t[i=l>=0?l--:o++])!=null&&!(2&c.__u)&&u==c.key&&p==c.type)return i}return-1}function de(e,t,n){t[0]=="-"?e.setProperty(t,n??""):e[t]=n==null?"":typeof n!="number"||it.test(t)?n:n+"px"}function V(e,t,n,s,l){var o,i;e:if(t=="style")if(typeof n=="string")e.style.cssText=n;else{if(typeof s=="string"&&(e.style.cssText=s=""),s)for(t in s)n&&t in n||de(e.style,t,"");if(n)for(t in n)s&&n[t]==s[t]||de(e.style,t,n[t])}else if(t[0]=="o"&&t[1]=="n")o=t!=(t=t.replace(ye,"$1")),i=t.toLowerCase(),t=i in e||t=="onFocusOut"||t=="onFocusIn"?i.slice(2):t.slice(2),e.l||(e.l={}),e.l[t+o]=n,n?s?n.u=s.u:(n.u=re,e.addEventListener(t,o?te:ee,o)):e.removeEventListener(t,o?te:ee,o);else{if(l=="http://www.w3.org/2000/svg")t=t.replace(/xlink(H|:h)/,"h").replace(/sName$/,"s");else if(t!="width"&&t!="height"&&t!="href"&&t!="list"&&t!="form"&&t!="tabIndex"&&t!="download"&&t!="rowSpan"&&t!="colSpan"&&t!="role"&&t!="popover"&&t in e)try{e[t]=n??"";break e}catch{}typeof n=="function"||(n==null||n===!1&&t[4]!="-"?e.removeAttribute(t):e.setAttribute(t,t=="popover"&&n==1?"":n))}}function me(e){return function(t){if(this.l){var n=this.l[t.type+e];if(t.t==null)t.t=re++;else if(t.t<n.u)return;return n(C.event?C.event(t):t)}}}function se(e,t,n,s,l,o,i,u,p,c){var d,a,h,_,f,v,y,m,g,S,k,R,x,P,D,L=t.type;if(t.constructor!==void 0)return null;128&n.__u&&(p=!!(32&n.__u),o=[u=t.__e=n.__e]),(d=C.__b)&&d(t);e:if(typeof L=="function")try{if(m=t.props,g=L.prototype&&L.prototype.render,S=(d=L.contextType)&&s[d.__c],k=d?S?S.props.value:d.__:s,n.__c?y=(a=t.__c=n.__c).__=a.__E:(g?t.__c=a=new L(m,k):(t.__c=a=new j(m,k),a.constructor=L,a.render=ft),S&&S.sub(a),a.state||(a.state={}),a.__n=s,h=a.__d=!0,a.__h=[],a._sb=[]),g&&a.__s==null&&(a.__s=a.state),g&&L.getDerivedStateFromProps!=null&&(a.__s==a.state&&(a.__s=F({},a.__s)),F(a.__s,L.getDerivedStateFromProps(m,a.__s))),_=a.props,f=a.state,a.__v=t,h)g&&L.getDerivedStateFromProps==null&&a.componentWillMount!=null&&a.componentWillMount(),g&&a.componentDidMount!=null&&a.__h.push(a.componentDidMount);else{if(g&&L.getDerivedStateFromProps==null&&m!==_&&a.componentWillReceiveProps!=null&&a.componentWillReceiveProps(m,k),t.__v==n.__v||!a.__e&&a.shouldComponentUpdate!=null&&a.shouldComponentUpdate(m,a.__s,k)===!1){t.__v!=n.__v&&(a.props=m,a.state=a.__s,a.__d=!1),t.__e=n.__e,t.__k=n.__k,t.__k.some(function(A){A&&(A.__=t)}),q.push.apply(a.__h,a._sb),a._sb=[],a.__h.length&&i.push(a);break e}a.componentWillUpdate!=null&&a.componentWillUpdate(m,a.__s,k),g&&a.componentDidUpdate!=null&&a.__h.push(function(){a.componentDidUpdate(_,f,v)})}if(a.context=k,a.props=m,a.__P=e,a.__e=!1,R=C.__r,x=0,g)a.state=a.__s,a.__d=!1,R&&R(t),d=a.render(a.props,a.state,a.context),q.push.apply(a.__h,a._sb),a._sb=[];else do a.__d=!1,R&&R(t),d=a.render(a.props,a.state,a.context),a.state=a.__s;while(a.__d&&++x<25);a.state=a.__s,a.getChildContext!=null&&(s=F(F({},s),a.getChildContext())),g&&!h&&a.getSnapshotBeforeUpdate!=null&&(v=a.getSnapshotBeforeUpdate(_,f)),P=d!=null&&d.type===$&&d.key==null?we(d.props.children):d,u=ke(e,z(P)?P:[P],t,n,s,l,o,i,u,p,c),a.base=t.__e,t.__u&=-161,a.__h.length&&i.push(a),y&&(a.__E=a.__=null)}catch(A){if(t.__v=null,p||o!=null)if(A.then){for(t.__u|=p?160:128;u&&u.nodeType==8&&u.nextSibling;)u=u.nextSibling;o[o.indexOf(u)]=null,t.__e=u}else{for(D=o.length;D--;)oe(o[D]);ne(t)}else t.__e=n.__e,t.__k=n.__k,A.then||ne(t);C.__e(A,t,n)}else o==null&&t.__v==n.__v?(t.__k=n.__k,t.__e=n.__e):u=t.__e=pt(n.__e,t,n,s,l,o,i,p,c);return(d=C.diffed)&&d(t),128&t.__u?void 0:u}function ne(e){e&&(e.__c&&(e.__c.__e=!0),e.__k&&e.__k.some(ne))}function Ce(e,t,n){for(var s=0;s<n.length;s++)ae(n[s],n[++s],n[++s]);C.__c&&C.__c(t,e),e.some(function(l){try{e=l.__h,l.__h=[],e.some(function(o){o.call(l)})}catch(o){C.__e(o,l.__v)}})}function we(e){return typeof e!="object"||e==null||e.__b>0?e:z(e)?e.map(we):F({},e)}function pt(e,t,n,s,l,o,i,u,p){var c,d,a,h,_,f,v,y=n.props||B,m=t.props,g=t.type;if(g=="svg"?l="http://www.w3.org/2000/svg":g=="math"?l="http://www.w3.org/1998/Math/MathML":l||(l="http://www.w3.org/1999/xhtml"),o!=null){for(c=0;c<o.length;c++)if((_=o[c])&&"setAttribute"in _==!!g&&(g?_.localName==g:_.nodeType==3)){e=_,o[c]=null;break}}if(e==null){if(g==null)return document.createTextNode(m);e=document.createElementNS(l,g,m.is&&m),u&&(C.__m&&C.__m(t,o),u=!1),o=null}if(g==null)y===m||u&&e.data==m||(e.data=m);else{if(o=o&&X.call(e.childNodes),!u&&o!=null)for(y={},c=0;c<e.attributes.length;c++)y[(_=e.attributes[c]).name]=_.value;for(c in y)_=y[c],c=="dangerouslySetInnerHTML"?a=_:c=="children"||c in m||c=="value"&&"defaultValue"in m||c=="checked"&&"defaultChecked"in m||V(e,c,null,_,l);for(c in m)_=m[c],c=="children"?h=_:c=="dangerouslySetInnerHTML"?d=_:c=="value"?f=_:c=="checked"?v=_:u&&typeof _!="function"||y[c]===_||V(e,c,_,y[c],l);if(d)u||a&&(d.__html==a.__html||d.__html==e.innerHTML)||(e.innerHTML=d.__html),t.__k=[];else if(a&&(e.innerHTML=""),ke(t.type=="template"?e.content:e,z(h)?h:[h],t,n,s,g=="foreignObject"?"http://www.w3.org/1999/xhtml":l,o,i,o?o[0]:n.__k&&O(n,0),u,p),o!=null)for(c=o.length;c--;)oe(o[c]);u||(c="value",g=="progress"&&f==null?e.removeAttribute("value"):f!=null&&(f!==e[c]||g=="progress"&&!f||g=="option"&&f!=y[c])&&V(e,c,f,y[c],l),c="checked",v!=null&&v!=e[c]&&V(e,c,v,y[c],l))}return e}function ae(e,t,n){try{if(typeof e=="function"){var s=typeof e.__u=="function";s&&e.__u(),s&&t==null||(e.__u=e(t))}else e.current=t}catch(l){C.__e(l,n)}}function xe(e,t,n){var s,l;if(C.unmount&&C.unmount(e),(s=e.ref)&&(s.current&&s.current!=e.__e||ae(s,null,t)),(s=e.__c)!=null){if(s.componentWillUnmount)try{s.componentWillUnmount()}catch(o){C.__e(o,t)}s.base=s.__P=null}if(s=e.__k)for(l=0;l<s.length;l++)s[l]&&xe(s[l],t,n||typeof e.type!="function");n||oe(e.__e),e.__c=e.__=e.__e=void 0}function ft(e,t,n){return this.constructor(e,n)}function Te(e,t,n){var s,l,o,i;t==document&&(t=document.documentElement),C.__&&C.__(e,t),l=(s=typeof n=="function")?null:n&&n.__k||t.__k,o=[],i=[],se(t,e=(!s&&n||t).__k=r($,null,[e]),l||B,B,t.namespaceURI,!s&&n?[n]:l?null:t.firstChild?X.call(t.childNodes):null,o,!s&&n?n:l?l.__e:t.firstChild,s,i),Ce(o,e,i)}X=q.slice,C={__e:function(e,t,n,s){for(var l,o,i;t=t.__;)if((l=t.__c)&&!l.__)try{if((o=l.constructor)&&o.getDerivedStateFromError!=null&&(l.setState(o.getDerivedStateFromError(e)),i=l.__d),l.componentDidCatch!=null&&(l.componentDidCatch(e,s||{}),i=l.__d),i)return l.__E=l}catch(u){e=u}throw e}},he=0,at=function(e){return e!=null&&e.constructor===void 0},j.prototype.setState=function(e,t){var n;n=this.__s!=null&&this.__s!=this.state?this.__s:this.__s=F({},this.state),typeof e=="function"&&(e=e(F({},n),this.props)),e&&F(n,e),e!=null&&this.__v&&(t&&this._sb.push(t),fe(this))},j.prototype.forceUpdate=function(e){this.__v&&(this.__e=!0,e&&this.__h.push(e),fe(this))},j.prototype.render=$,H=[],ve=typeof Promise=="function"?Promise.prototype.then.bind(Promise.resolve()):setTimeout,ge=function(e,t){return e.__v.__b-t.__v.__b},J.__r=0,ye=/(PointerCapture)$|Capture$/i,re=0,ee=me(!1),te=me(!0),lt=0;var M,w,le,Pe,K=0,Ue=[],T=C,Ee=T.__b,Re=T.__r,De=T.diffed,Le=T.__c,Ie=T.unmount,Fe=T.__;function ce(e,t){T.__h&&T.__h(w,e,K||t),K=0;var n=w.__H||(w.__H={__:[],__h:[]});return e>=n.__.length&&n.__.push({}),n.__[e]}function b(e){return K=1,dt(Ae,e)}function dt(e,t,n){var s=ce(M++,2);if(s.t=e,!s.__c&&(s.__=[n?n(t):Ae(void 0,t),function(u){var p=s.__N?s.__N[0]:s.__[0],c=s.t(p,u);p!==c&&(s.__N=[c,s.__[1]],s.__c.setState({}))}],s.__c=w,!w.__f)){var l=function(u,p,c){if(!s.__c.__H)return!0;var d=s.__c.__H.__.filter(function(h){return h.__c});if(d.every(function(h){return!h.__N}))return!o||o.call(this,u,p,c);var a=s.__c.props!==u;return d.some(function(h){if(h.__N){var _=h.__[0];h.__=h.__N,h.__N=void 0,_!==h.__[0]&&(a=!0)}}),o&&o.call(this,u,p,c)||a};w.__f=!0;var o=w.shouldComponentUpdate,i=w.componentWillUpdate;w.componentWillUpdate=function(u,p,c){if(this.__e){var d=o;o=void 0,l(u,p,c),o=d}i&&i.call(this,u,p,c)},w.shouldComponentUpdate=l}return s.__N||s.__}function I(e,t){var n=ce(M++,3);!T.__s&&Ne(n.__H,t)&&(n.__=e,n.u=t,w.__H.__h.push(n))}function N(e){return K=5,mt(function(){return{current:e}},[])}function mt(e,t){var n=ce(M++,7);return Ne(n.__H,t)&&(n.__=e(),n.__H=t,n.__h=e),n.__}function ht(){for(var e;e=Ue.shift();){var t=e.__H;if(e.__P&&t)try{t.__h.some(G),t.__h.some(ie),t.__h=[]}catch(n){t.__h=[],T.__e(n,e.__v)}}}T.__b=function(e){w=null,Ee&&Ee(e)},T.__=function(e,t){e&&t.__k&&t.__k.__m&&(e.__m=t.__k.__m),Fe&&Fe(e,t)},T.__r=function(e){Re&&Re(e),M=0;var t=(w=e.__c).__H;t&&(le===w?(t.__h=[],w.__h=[],t.__.some(function(n){n.__N&&(n.__=n.__N),n.u=n.__N=void 0})):(t.__h.some(G),t.__h.some(ie),t.__h=[],M=0)),le=w},T.diffed=function(e){De&&De(e);var t=e.__c;t&&t.__H&&(t.__H.__h.length&&(Ue.push(t)!==1&&Pe===T.requestAnimationFrame||((Pe=T.requestAnimationFrame)||vt)(ht)),t.__H.__.some(function(n){n.u&&(n.__H=n.u),n.u=void 0})),le=w=null},T.__c=function(e,t){t.some(function(n){try{n.__h.some(G),n.__h=n.__h.filter(function(s){return!s.__||ie(s)})}catch(s){t.some(function(l){l.__h&&(l.__h=[])}),t=[],T.__e(s,n.__v)}}),Le&&Le(e,t)},T.unmount=function(e){Ie&&Ie(e);var t,n=e.__c;n&&n.__H&&(n.__H.__.some(function(s){try{G(s)}catch(l){t=l}}),n.__H=void 0,t&&T.__e(t,n.__v))};var He=typeof requestAnimationFrame=="function";function vt(e){var t,n=function(){clearTimeout(s),He&&cancelAnimationFrame(t),setTimeout(e)},s=setTimeout(n,35);He&&(t=requestAnimationFrame(n))}function G(e){var t=w,n=e.__c;typeof n=="function"&&(e.__c=void 0,n()),w=t}function ie(e){var t=w;e.__c=e.__(),w=t}function Ne(e,t){return!e||e.length!==t.length||t.some(function(n,s){return n!==e[s]})}function Ae(e,t){return typeof t=="function"?t(e):t}var Q=null;async function E(e,t={}){let n=await fetch(e,t);if(!n.ok){let s=await n.json().catch(()=>({error:n.statusText}));throw new Error(s.error||n.statusText)}return n.json()}async function U(){Q||(Q=(await E("/api/v1/csrf")).token);let e=Q;return Q=null,e}async function ue(){return E("/api/v1/status")}async function Oe(){return E("/api/v1/rules")}async function $e(e=""){return E(`/api/v1/secrets/${e}`)}async function Y(e,t=!1){let n=t?`/api/v1/secrets/${e}?reveal=true`:`/api/v1/secrets/${e}`;return E(n)}async function Me(){let e=await U();return E("/api/v1/sync",{method:"POST",headers:{"X-CSRF-Token":e}})}async function Ve(e,t){let n=await U();return E("/auth/ldap/login",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":n},body:JSON.stringify({username:e,password:t})})}async function We(e){return E(`/auth/ldap/status?session=${encodeURIComponent(e)}`)}async function je(e,t){let n=await U();return E("/auth/ldap/totp",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":n},body:JSON.stringify({session_id:e,passcode:t})})}async function Be(e){let t=await U();return E("/auth/token/login",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":t},body:JSON.stringify({token:e})})}async function qe(){return E("/api/v1/token")}async function Je(e){let t=await U();return E(`/api/v1/enrol/${encodeURIComponent(e)}/start`,{method:"POST",headers:{"X-CSRF-Token":t}})}async function Xe(e){let t=await U();return E(`/api/v1/enrol/${encodeURIComponent(e)}/skip`,{method:"POST",headers:{"X-CSRF-Token":t}})}async function ze(e){return E(`/api/v1/enrol/${encodeURIComponent(e)}/status`)}async function Ge(){let e=await U();return E("/api/v1/enrol/complete",{method:"POST",headers:{"X-CSRF-Token":e}})}async function Ke(){return E("/api/v1/enrol/prompt")}async function Qe(e){let t=await U();return E("/api/v1/enrol/secret",{method:"POST",headers:{"Content-Type":"application/json","X-CSRF-Token":t},body:JSON.stringify({value:e})})}function Z(e){let t=document.createElement("textarea");t.value=e,t.style.position="fixed",t.style.opacity="0",document.body.appendChild(t),t.select(),document.execCommand("copy"),document.body.removeChild(t)}function Ye({status:e,onSync:t,pendingEnrolments:n,onEnrolClick:s}){let[l,o]=b(!1),[i,u]=b(!1),p=N(null);async function c(){o(!0);try{await Me(),t&&await t()}catch(v){console.error("sync failed:",v)}finally{o(!1)}}async function d(){try{if(!p.current){let v=await qe();p.current=v.token}Z(p.current),u(!0),setTimeout(()=>u(!1),2e3)}catch(v){console.error("copy token failed:",v)}}let a=e?.authenticated?"Connected":"Disconnected",h=e?.authenticated?"status-ok":"status-error",_=e?.vault_address,f=_&&/^https?:\/\//i.test(_)?_:null;return r("header",{class:"status-bar"},r("div",{class:"status-left"},r("span",{class:"app-title"},".vault",e?.version&&r("span",{class:"app-version"}," v"+e.version)),r("span",{class:`status-indicator ${h}`},a),f&&r("a",{class:"vault-link",href:f,target:"_blank",rel:"noopener noreferrer"},"Vault \u2197"),n>0&&r("button",{class:"enrol-indicator",onClick:s},n+" pending")),r("div",{class:"status-right"},e?.time&&r("span",{class:"last-sync"},"Updated: ",new Date(e.time).toLocaleTimeString()),e?.authenticated&&r("button",{class:"copy-token-btn"+(i?" copied":""),onClick:d,title:i?"Token copied!":"Copy Vault token to clipboard"},i?"\u2705":"\u{1F4CB}"),r("button",{class:"sync-btn",onClick:c,disabled:l},l?"Syncing...":"Sync Now")))}function Ze({keys:e,selected:t,onSelect:n}){return!e||e.length===0?r("nav",{class:"sidebar"},r("div",{class:"sidebar-empty"},"No secrets found")):r("nav",{class:"sidebar"},r("div",{class:"sidebar-header"},"Secrets"),r("ul",{class:"key-list"},e.map(s=>r("li",{key:s,class:`key-item ${s===t?"selected":""}`,onClick:()=>n(s)},r("span",{class:"key-icon"},s.endsWith("/")?"\u{1F4C1}":"\u{1F511}"),r("span",{class:"key-name"},s)))))}function gt(e){return!!e&&/^https?:\/\//i.test(e)}function yt(e,t){if(!gt(e.vault_address))return null;let n=e.vault_address.replace(/\/+$/,""),s=t&&t.endsWith("/")?"list":"show",l=encodeURIComponent(e.kv_mount),o=c=>(c||"").split("/").filter(d=>d.length>0).map(d=>encodeURIComponent(d)).join("/"),i=o(`${e.user_prefix||""}${e.username}`),u=o(t),p=u?`${i}/${u}`:i;return`${n}/ui/vault/secrets/${l}/${s}/${p}`}function et({secretPath:e,status:t,customText:n}){let[s,l]=b(null),[o,i]=b({}),[u,p]=b({}),[c,d]=b(!1);I(()=>{if(!e){l(null);return}d(!0),i({}),Y(e).then(l).catch(()=>l(null)).finally(()=>d(!1))},[e]);async function a(f){if(o[f]){i(y=>{let m={...y};return delete m[f],m});return}let v=await Y(e,!0);v&&v.fields&&(i(y=>({...y,[f]:v.fields[f]})),setTimeout(()=>{i(y=>{let m={...y};return delete m[f],m})},3e4))}async function h(f){let v=await Y(e,!0);if(v&&v.fields){let y=typeof v.fields[f]=="object"?JSON.stringify(v.fields[f],null,2):String(v.fields[f]);await navigator.clipboard.writeText(y),p(m=>({...m,[f]:!0})),setTimeout(()=>{p(m=>{let g={...m};return delete g[f],g})},2e3)}}if(!e)return r("main",{class:"secret-panel"},n?r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:n}}):r("div",{class:"panel-empty"},"Select a secret from the sidebar"));if(c)return r("main",{class:"secret-panel"},r("div",{class:"panel-loading"},"Loading..."));if(!s)return r("main",{class:"secret-panel"},r("div",{class:"panel-error"},"Secret not found"));let _=Array.isArray(s.fields)?s.fields:Object.keys(s.fields||{});return r("main",{class:"secret-panel"},r("div",{class:"secret-heading"},r("h2",null,e),(()=>{let f=t?.kv_mount&&t?.username&&yt(t,e);return f&&r("a",{class:"vault-link",href:f,target:"_blank",rel:"noopener noreferrer"},"View in Vault \u2197")})()),r("div",{class:"secret-meta"},"Version: ",s.version),r("table",{class:"field-table"},r("thead",null,r("tr",null,r("th",null,"Field"),r("th",null,"Value"),r("th",null,""))),r("tbody",null,_.map(f=>r("tr",{key:f},r("td",{class:"field-name"},f),r("td",{class:"field-value"},o[f]?r("code",null,typeof o[f]=="object"?JSON.stringify(o[f],null,2):String(o[f])):r("span",{class:"masked"},"\u2022\u2022\u2022\u2022\u2022\u2022\u2022\u2022")),r("td",{class:"field-actions"},r("button",{class:"reveal-btn",onClick:()=>a(f),title:o[f]?"Hide":"Reveal"},o[f]?"\u{1F441}\uFE0F":"\u{1F441}"),r("button",{class:"copy-btn"+(u[f]?" copied":""),onClick:()=>h(f),title:u[f]?"Copied!":"Copy to clipboard"},u[f]?"\u2705":"\u{1F4CB}")))))))}function tt({rules:e}){return!e||e.length===0?null:r("div",{class:"oauth-banners"},e.map(t=>r("div",{key:t.name,class:"oauth-banner"},r("span",null,t.oauth_provider||t.name," requires authorization"),r("a",{class:"oauth-btn",href:`/api/v1/oauth/${t.name}/start`},"Authorize"))))}function nt({authMethod:e,onAuth:t,customText:n}){return e==="oidc"?r(bt,{customText:n}):e==="ldap"?r(kt,{onAuth:t,customText:n}):e==="token"?r(St,{onAuth:t,customText:n}):r("div",{class:"login-container"},r("p",{class:"login-error"},`Unknown auth method: ${e}`))}function bt({customText:e}){return r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),e&&r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:e}}),r("a",{class:"login-btn",href:"/auth/oidc/start"},"Login with OIDC")))}function kt({onAuth:e,customText:t}){let[n,s]=b(""),[l,o]=b(""),[i,u]=b(null),[p,c]=b("form"),[d,a]=b([]),[h,_]=b(""),f=N(null),v=N(null);I(()=>()=>{v.current&&clearInterval(v.current)},[]);function y(S){v.current&&clearInterval(v.current),v.current=setInterval(async()=>{try{let k=await We(S);if(k.state==="mfa_required"){a(k.mfa_methods||[]);let R=k.mfa_methods&&k.mfa_methods.some(x=>x.uses_passcode);c(R?"totp":"polling"),k.error&&u(k.error)}else k.state==="authenticated"?(clearInterval(v.current),e&&e()):k.state==="failed"&&(clearInterval(v.current),u(k.error||"Login failed"),c("form"))}catch(k){clearInterval(v.current),u(k.message),c("form")}},2e3)}async function m(S){S.preventDefault(),u(null),c("polling");try{let k=await Ve(n,l);f.current=k.session_id,y(k.session_id)}catch(k){u(k.message),c("form")}}async function g(S){S.preventDefault(),u(null);try{await je(f.current,h),_("")}catch(k){u(k.message)}}return p==="polling"&&(!d.length||!d.some(S=>S.uses_passcode))?r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),r("p",{class:"login-mfa-wait"},"Waiting for MFA approval..."),r("p",{class:"login-mfa-hint"},"Check your device"),i&&r("p",{class:"login-error"},i))):p==="totp"?r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),r("p",{class:"login-subtitle"},"Enter MFA passcode"),i&&r("p",{class:"login-error"},i),r("form",{onSubmit:g},r("input",{type:"text",class:"login-input",placeholder:"Passcode","aria-label":"MFA Passcode",value:h,onInput:S=>_(S.target.value),autocomplete:"one-time-code",inputMode:"numeric",autofocus:!0}),r("button",{type:"submit",class:"login-btn",disabled:!h},"Verify")))):r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),t&&r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:t}}),i&&r("p",{class:"login-error"},i),r("form",{onSubmit:m},r("input",{type:"text",class:"login-input",placeholder:"Username","aria-label":"Username",value:n,onInput:S=>s(S.target.value),autocomplete:"username",autofocus:!0}),r("input",{type:"password",class:"login-input",placeholder:"Password","aria-label":"Password",value:l,onInput:S=>o(S.target.value),autocomplete:"current-password"}),r("button",{type:"submit",class:"login-btn",disabled:!n||!l},"Login"))))}function St({onAuth:e,customText:t}){let[n,s]=b(""),[l,o]=b(null),[i,u]=b(!1);async function p(c){c.preventDefault(),o(null),u(!0);try{await Be(n),e&&e()}catch(d){o(d.message)}finally{u(!1)}}return r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),t&&r("div",{class:"custom-text",dangerouslySetInnerHTML:{__html:t}}),l&&r("p",{class:"login-error"},l),r("form",{onSubmit:p},r("input",{type:"password",class:"login-input",placeholder:"Vault Token","aria-label":"Vault Token",value:n,onInput:c=>s(c.target.value),autofocus:!0}),r("button",{type:"submit",class:"login-btn",disabled:!n||i},i?"Validating...":"Login"))))}function rt({enrolment:e,onUpdate:t,anyRunning:n}){let[s,l]=b(e.status),[o,i]=b(e.output||[]),[u,p]=b(e.error||null),[c,d]=b(null),[a,h]=b(""),_=N(null);I(()=>{l(e.status),p(e.error||null),e.output&&i(e.output)},[e.status,e.error,e.output]),I(()=>(s==="running"&&!_.current&&f(),()=>{_.current&&(clearInterval(_.current),_.current=null)}),[s]);function f(){_.current&&(clearInterval(_.current),_.current=null);let x=e.engine!=="github";_.current=setInterval(async()=>{try{let P=await ze(e.key);if(i(P.output||[]),x){let D=await Ke();D.pending?d(D.label):d(null)}P.status!=="running"&&(clearInterval(_.current),_.current=null,l(P.status),p(P.error||null),d(null),t&&t())}catch(P){console.error("poll error:",P)}},2e3)}async function v(){try{await Je(e.key),l("running"),i([]),p(null),t&&t(),f()}catch(x){p(x.message)}}async function y(){try{await Xe(e.key),l("skipped"),t&&t()}catch(x){p(x.message)}}async function m(x){x.preventDefault();try{await Qe(a),h(""),d(null)}catch(P){p(P.message)}}let g=o.reduce((x,P)=>{let D=P.match(/one-time code: (\S+)/);return D?D[1]:x},null),S=o.reduce((x,P)=>{let D=P.match(/open (https?:\/\/\S+)/);return D?D[1]:x},null),k=e.engine==="github",R=n&&s!=="running";return s==="complete"?r("div",{class:"enrol-card enrol-complete"},r("div",{class:"enrol-card-header"},r("div",null,r("span",{class:"enrol-check"},"\u2713"),r("strong",null,e.name)),r("span",{class:"enrol-status-text enrol-status-complete"},"Enrolled successfully"))):s==="skipped"?r("div",{class:"enrol-card enrol-skipped"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-badge"},"SKIPPED")),r("span",{class:"enrol-engine-desc"},_e(e.engine)))):s==="running"?r("div",{class:"enrol-card enrol-running"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-badge enrol-badge-running"},"RUNNING"))),k&&g&&r("div",{class:"enrol-device-flow"},r("p",{class:"enrol-device-label"},"Enter this code on GitHub:"),r("div",{class:"enrol-device-code"},g),r("div",{class:"enrol-device-actions"},r("button",{class:"enrol-btn-secondary",onClick:()=>Z(g)},"Copy Code"),S&&r("a",{class:"enrol-btn-secondary",href:S,target:"_blank",rel:"noopener noreferrer"},"Open GitHub \u2192")),r("p",{class:"enrol-device-waiting"},"Waiting for approval...")),c&&!k&&r("form",{class:"enrol-prompt-form",onSubmit:m},r("label",{class:"enrol-prompt-label"},c),r("input",{type:"password",class:"enrol-prompt-input",value:a,onInput:x=>h(x.target.value),placeholder:"Enter passphrase",autofocus:!0}),r("div",{class:"enrol-prompt-actions"},r("button",{type:"submit",class:"enrol-btn-primary"},"Submit"))),!k&&!c&&o.length>0&&r("div",{class:"enrol-output"},o.map((x,P)=>r("div",{key:P},x)))):s==="failed"?r("div",{class:"enrol-card enrol-failed"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-engine-desc"},_e(e.engine))),r("div",{class:"enrol-card-actions"},r("button",{class:"enrol-btn-primary",onClick:v,disabled:R},"Retry"),r("button",{class:"enrol-btn-secondary",onClick:y,disabled:R},"Skip"))),r("p",{class:"enrol-error-text"},u)):r("div",{class:"enrol-card"},r("div",{class:"enrol-card-header"},r("div",null,r("strong",null,e.name),r("span",{class:"enrol-engine-desc"},_e(e.engine))),r("div",{class:"enrol-card-actions"},r("button",{class:"enrol-btn-primary",onClick:v,disabled:R},"Start"),r("button",{class:"enrol-btn-secondary",onClick:y,disabled:R},"Skip"))),u&&r("p",{class:"enrol-error-text"},u))}function _e(e){switch(e){case"github":return"OAuth token via device flow";case"ssh":return"Ed25519 key generation";default:return e}}function ot({enrolments:e,onComplete:t,onUpdate:n}){let s=e.every(i=>i.status==="complete"||i.status==="skipped"),l=e.some(i=>i.status==="running");async function o(){try{await Ge(),t&&t()}catch(i){console.error("complete error:",i)}}return r("div",{class:"enrol-page"},r("div",{class:"enrol-container"},r("h2",{class:"enrol-heading"},"Complete your setup"),r("p",{class:"enrol-subheading"},"The following credentials need to be configured before syncing can begin."),r("div",{class:"enrol-list"},e.map(i=>r(rt,{key:i.key,enrolment:i,onUpdate:n,anyRunning:l}))),r("div",{class:"enrol-footer"},r("button",{class:"enrol-continue-btn",onClick:o,disabled:l},s?"Continue to Dashboard \u2192":"Skip remaining and continue \u2192"))))}function st(){let[e,t]=b(null),[n,s]=b([]),[l,o]=b([]),[i,u]=b(null),[p,c]=b(null),[d,a]=b(!1);I(()=>{h()},[]),I(()=>{if(!e?.authenticated)return;let m=setInterval(_,3e4);return()=>clearInterval(m)},[e?.authenticated]);async function h(){try{let m=await ue();t(m),m.authenticated&&_()}catch(m){c(m.message)}}async function _(){try{let[m,g,S]=await Promise.all([ue(),Oe(),$e()]);t(m),s(g.rules||[]),o(S.keys||[]),c(null)}catch(m){c(m.message)}}if(e&&!e.authenticated)return r(nt,{authMethod:e.auth_method,onAuth:_,customText:e.login_text});if(!e)return r("div",{class:"login-container"},r("div",{class:"login-card"},r("h1",{class:"login-title"},".vault"),p?r("p",{class:"login-error"},p):r("p",null,"Loading...")));let f=e.enrolments||[],v=f.filter(m=>m.status==="pending"||m.status==="running"||m.status==="failed");if(v.length>0&&!d)return r(ot,{enrolments:f,onComplete:()=>{a(!0),_()},onUpdate:h});let y=n.filter(m=>m.has_oauth);return r($,null,r(Ye,{status:e,onSync:_,pendingEnrolments:v.length,onEnrolClick:()=>a(!1)}),p&&r("div",{class:"error-banner"},p),y.length>0&&r(tt,{rules:y}),r("div",{class:"main-layout"},r(Ze,{keys:l,selected:i,onSelect:u}),r(et,{secretPath:i,status:e,customText:e.secret_view_text})))}Te(r(st,null),document.getElementById("app"));})();
